+++ server/jobs/send-users.js
@@ -29,7 +29,7 @@ function sendUsers(ctx, payload) {
             u["traits_intercom/id"] = intercomData.id;
             u["traits_intercom/tags"] = intercomData.tags.tags.map(t => t.name);
 
-            ctx.client.asUser(_.pick(u, ["email", "id", "external_id"])).logger.info("outgoing.user.success");
+            ctx.client.asUser(_.pick(u, ["email", "id", "external_id"])).logger.info("outgoing.user.success", intercomData);
             return u;
           });
         const errors = _.filter(res, { body: { type: "error.list" } });